pyproject.toml
[tool.poetry]
name = "db-migration"
version = "0.0.1"
description = "This is simple script to migrate your data from one db to another."
authors = ["Mohammad Ashraful <ashrafulrobin3@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "db_migration"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
pandas = "^1.5.2"
pymssql = "^2.2.7"
sqlalchemy = "^1.4.45"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
